re-structuring project and added constants to end point
