+++ app/src/main/java/mx/kodemia/personalweather/core/Constants.kt
@@ -2,6 +2,11 @@ package mx.kodemia.personalweather.core
 
 object Constants {
     const val BASE_URL = "https://api.openweathermap.org/"
+    const val API_WEATHER_END_POINT = "data/2.5/onecall"
+    const val API_CITY_END_POINT = "geo/1.0/reverse"
     const val API_KEY = "804c1de62c103c56406e6fe94bf50b86"
+    const val EXCLUDE_PARAMETERS = "minutely,hourly,alerts"
+
     const val REQUEST_PERMISSIONS_REQUEST_CODE = 34
+
 }
