--- app/src/main/java/mx/kodemia/personalweather/core/utils/LoadingState.kt
@@ -1,6 +1,0 @@
-package mx.kodemia.personalweather.core.utils
-
-sealed class LoadingState{
-    data class Loading(val loading: Boolean = true)
-    data class Error(val error: String)
-}
