+++ app/src/main/java/mx/kodemia/personalweather/data/model/city/CityEntity.kt
@@ -1,0 +1,8 @@
+package mx.kodemia.personalweather.data.model.city
+
+import mx.kodemia.personalweather.data.model.weather.Current
+
+data class CityEntity(
+    val current: Current,
+    var city: City?
+)
