+++ app/src/main/java/mx/kodemia/personalweather/data/model/weather/Main.kt
@@ -1,4 +1,4 @@
-package mx.kodemia.personalweather.model.weather
+package mx.kodemia.personalweather.data.model.weather
 
 data class Main(
     val temp: Double,
