+++ app/src/main/java/mx/kodemia/personalweather/data/model/weather/Wind.kt
@@ -1,0 +1,5 @@
+package mx.kodemia.personalweather.data.model.weather
+
+data class Wind(
+    val speed: Double,
+)
