+++ app/src/main/java/mx/kodemia/personalweather/data/model/weather_daily/Daily.kt
@@ -1,4 +1,4 @@
-package mx.kodemia.personalweather.model.weather_daily
+package mx.kodemia.personalweather.data.model.weather_daily
 
 data class Daily(
     val daily: ArrayList<WeatherDaily>
