+++ app/src/main/java/mx/kodemia/personalweather/data/model/weather_daily/DailyTemp.kt
@@ -1,0 +1,5 @@
+package mx.kodemia.personalweather.data.model.weather_daily
+
+data class DailyTemp(
+    val day: Double
+)
