+++ app/src/main/java/mx/kodemia/personalweather/data/model/weather_daily/WeatherDaily.kt
@@ -1,6 +1,6 @@
-package mx.kodemia.personalweather.model.weather_daily
+package mx.kodemia.personalweather.data.model.weather_daily
 
-import mx.kodemia.personalweather.model.weather.Weather
+import mx.kodemia.personalweather.data.model.weather.Weather
 
 data class WeatherDaily(
     val dt: Long,
