--- app/src/main/java/mx/kodemia/personalweather/model/city/CityEntity.kt
@@ -1,8 +1,0 @@
-package mx.kodemia.personalweather.model.city
-
-import mx.kodemia.personalweather.model.weather.Current
-
-data class CityEntity(
-    val current: Current,
-    var city: City?
-)
