--- app/src/main/java/mx/kodemia/personalweather/model/weather/Wind.kt
@@ -1,5 +1,0 @@
-package mx.kodemia.personalweather.model.weather
-
-data class Wind(
-    val speed: Double,
-)
