--- app/src/main/java/mx/kodemia/personalweather/model/weather_daily/DailyTemp.kt
@@ -1,5 +1,0 @@
-package mx.kodemia.personalweather.model.weather_daily
-
-data class DailyTemp(
-    val day: Double
-)
